Fixed error with carts not disappearing
This error did not appear until after the error with the enemy box
colliders were fixed. The cars would no longer disappear because the
colliders covered the area of each car, but it didn't include the hero
or bystanders. The clown are programmed to track the camera and to
self-destruct when they hit the camera (centered on the hero), but since
they never hit the cars, they wouldn't trigger the action to cause the
cars to disappear.  I kind of cheated the solution by extending the
front cart's collider. I'll see about getting a better solution worked
out soon.