Back of track now extendable with a variable
The back of the track will now extend backwards from the position of the
player a distance denoted as backLength that is specified in the
GenerateRails script. The code for moving the rails should be more
optimized now, as it only moves the rails it needs to move (the back
rail to the front). In fact, in it's present state, no rails are moving
because the track isn't changing....and we aren't really moving...things
to think about.